\name{GreedyEPL-package}
\alias{GreedyEPL-package}
\alias{GreedyEPL}
\docType{package}
\title{
\packageTitle{GreedyEPL}
}
\description{
\packageDescription{GreedyEPL}
}
\author{
Riccardo Rastelli

Mantainer: Riccardo Rastelli <riccardo.rastelli@wu.ac.at>
}
\references{
Rastelli, R. and Friel, N. (2017) "Optimal Bayesian estimators for latent variable cluster models" <DOI:10.1007/s11222-017-9786-y>
}
\keyword{ package }