\name{GreedyEPL-package}
\alias{GreedyEPL-package}
\alias{GreedyEPL}
\docType{package}
\title{
\packageTitle{GreedyEPL}
}
\description{
\packageDescription{GreedyEPL}
}
\author{
Riccardo Rastelli

Mantainer: Riccardo Rastelli <riccardo.rastelli@wu.ac.at>
}
\references{
Rastelli, R. and Friel, N. (2016) "Optimal Bayesian estimators for latent variable cluster models", \url{https://arxiv.org/abs/1607.02325}
}
\keyword{ package }